Backend/requirements.txt
asgiref==3.11.0
django==6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
pyjwt==2.10.1
python-decouple==3.8
pytz==2025.2
pyyaml==6.0.3
sqlparse==0.5.4
tzdata==2025.3
uritemplate==4.2.0
pip freeze > requirements.txt
